# contracts/agent-api.openapi.json
{"openapi":"3.1.0","info":{"title":"WoodHub AI Agent","version":"1.0"},"paths":{"/v1/agent/chat":{"post":{"tags":["agent"],"summary":"Customer Chat","operationId":"customer_chat_v1_agent_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgentResponse"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Unprocessable Entity","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"429":{"description":"Too Many Requests","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"503":{"description":"Service Unavailable","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/v1/agent/chat/stream":{"post":{"tags":["agent"],"summary":"Customer Chat Stream","operationId":"customer_chat_stream_v1_agent_chat_stream_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Unprocessable Entity","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"429":{"description":"Too Many Requests","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"503":{"description":"Service Unavailable","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/v1/agent/manage/chat":{"post":{"tags":["agent"],"summary":"Manage Chat","operationId":"manage_chat_v1_agent_manage_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgentResponse"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Unprocessable Entity","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"429":{"description":"Too Many Requests","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"503":{"description":"Service Unavailable","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/v1/agent/manage/chat/stream":{"post":{"tags":["agent"],"summary":"Manage Chat Stream","operationId":"manage_chat_stream_v1_agent_manage_chat_stream_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"403":{"description":"Forbidden","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Unprocessable Entity","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"429":{"description":"Too Many Requests","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"503":{"description":"Service Unavailable","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/v1/agent/actions/{action_id}":{"get":{"tags":["agent"],"summary":"Get Action","operationId":"get_action_v1_agent_actions__action_id__get","parameters":[{"name":"action_id","in":"path","required":true,"schema":{"type":"string","title":"Action Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActionOut"}}}},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Unauthorized"},"403":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Forbidden"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Unprocessable Entity"},"429":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Too Many Requests"},"503":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Service Unavailable"}}}},"/v1/agent/actions/{action_id}/confirm":{"post":{"tags":["agent"],"summary":"Confirm Action","operationId":"confirm_action_v1_agent_actions__action_id__confirm_post","parameters":[{"name":"action_id","in":"path","required":true,"schema":{"type":"string","title":"Action Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgentResponse"}}}},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Unauthorized"},"403":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Forbidden"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Unprocessable Entity"},"429":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Too Many Requests"},"503":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Service Unavailable"}}}},"/v1/agent/actions/{action_id}/cancel":{"post":{"tags":["agent"],"summary":"Cancel Action","operationId":"cancel_action_v1_agent_actions__action_id__cancel_post","parameters":[{"name":"action_id","in":"path","required":true,"schema":{"type":"string","title":"Action Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgentResponse"}}}},"401":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Unauthorized"},"403":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Forbidden"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Unprocessable Entity"},"429":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Too Many Requests"},"503":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}},"description":"Service Unavailable"}}}},"/v1/agent/capabilities":{"get":{"tags":["agent","agent"],"summary":"Capabilities","description":"Danh sách tool người gọi được dùng (để Frontend hiển thị gợi ý). Không chứa dữ liệu nhạy cảm.","operationId":"capabilities_v1_agent_capabilities_get","parameters":[{"name":"profile","in":"query","required":false,"schema":{"$ref":"#/components/schemas/AgentProfile","default":"customer"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Capabilities V1 Agent Capabilities Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chat":{"post":{"tags":["legacy"],"summary":"Legacy Chat","operationId":"legacy_chat_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LegacyChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"deprecated":true}},"/health":{"get":{"tags":["monitoring"],"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Health Get"}}}}}}}},"components":{"schemas":{"ActionOut":{"properties":{"id":{"type":"string","title":"Id"},"tool":{"type":"string","title":"Tool"},"operation":{"type":"string","enum":["UPDATE","SENSITIVE_UPDATE","ACTION"],"title":"Operation"},"state":{"type":"string","enum":["pending_confirmation","confirmed","executing","verified","completed","unverified","failed","cancelled","expired"],"title":"State"},"summary":{"type":"string","title":"Summary"},"target":{"additionalProperties":{"type":"string"},"type":"object","title":"Target"},"changes":{"items":{"$ref":"#/components/schemas/FieldChangeOut"},"type":"array","title":"Changes"},"warnings":{"items":{"type":"string"},"type":"array","title":"Warnings"},"confirmation":{"anyOf":[{"$ref":"#/components/schemas/ConfirmationOut"},{"type":"null"}]},"verified":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Verified"},"error_code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error Code"},"error_message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error Message"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["id","tool","operation","state","summary","target","changes","created_at","updated_at"],"title":"ActionOut"},"AgentProfile":{"type":"string","enum":["customer","management"],"title":"AgentProfile"},"AgentResponse":{"properties":{"type":{"type":"string","enum":["answer","clarification","confirmation_required","action_result","error"],"title":"Type"},"message":{"type":"string","title":"Message"},"session_id":{"type":"string","title":"Session Id"},"request_id":{"type":"string","title":"Request Id"},"blocks":{"items":{"$ref":"#/components/schemas/Block"},"type":"array","title":"Blocks"},"sources":{"items":{"$ref":"#/components/schemas/SourceOut"},"type":"array","title":"Sources"},"action":{"anyOf":[{"$ref":"#/components/schemas/ActionOut"},{"type":"null"}]},"error":{"anyOf":[{"$ref":"#/components/schemas/ErrorOut"},{"type":"null"}]},"meta":{"$ref":"#/components/schemas/MetaOut"}},"type":"object","required":["type","message","session_id","request_id","meta"],"title":"AgentResponse"},"Block":{"properties":{"kind":{"type":"string","enum":["product_list","product_detail","product_comparison","inventory","store_info","branch_list","workshop_list","promotion_list","policy","knowledge","taxonomy","design_task","candidates"],"title":"Kind"},"data":{"title":"Data"}},"type":"object","required":["kind","data"],"title":"Block"},"ChatRequest":{"properties":{"message":{"type":"string","maxLength":10000,"minLength":1,"title":"Message","description":"Tin nhắn người dùng (giới hạn thực tế theo MAX_MESSAGE_CHARS)."},"session_id":{"anyOf":[{"type":"string","maxLength":100,"pattern":"^[A-Za-z0-9_\\-:.]+$"},{"type":"null"}],"title":"Session Id","description":"Bỏ trống để server tạo phiên mới."},"client_message_id":{"anyOf":[{"type":"string","maxLength":100},{"type":"null"}],"title":"Client Message Id","description":"Id phía client để chống gửi trùng."},"location":{"anyOf":[{"$ref":"#/components/schemas/Location"},{"type":"null"}],"description":"Vị trí thiết bị (chỉ dùng cho tìm xưởng gần)."}},"type":"object","required":["message"],"title":"ChatRequest"},"ConfirmRequest":{"properties":{"confirmation_code":{"type":"string","maxLength":12,"minLength":4,"title":"Confirmation Code"},"session_id":{"anyOf":[{"type":"string","maxLength":100},{"type":"null"}],"title":"Session Id"}},"type":"object","required":["confirmation_code"],"title":"ConfirmRequest"},"ConfirmationOut":{"properties":{"required":{"type":"boolean","title":"Required"},"level":{"type":"string","enum":["standard","strong"],"title":"Level"},"code":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Code","description":"Mã người dùng phải gửi lại để xác nhận (chỉ trả cho chủ action)."},"expires_at":{"type":"string","format":"date-time","title":"Expires At"},"confirm_endpoint":{"type":"string","title":"Confirm Endpoint"},"cancel_endpoint":{"type":"string","title":"Cancel Endpoint"},"chat_phrase":{"type":"string","title":"Chat Phrase","description":"Câu người dùng có thể gõ trong chat để xác nhận, vd 'xác nhận ABC123'."}},"type":"object","required":["required","level","expires_at","confirm_endpoint","cancel_endpoint","chat_phrase"],"title":"ConfirmationOut"},"ErrorOut":{"properties":{"code":{"type":"string","title":"Code"},"message":{"type":"string","title":"Message"}},"type":"object","required":["code","message"],"title":"ErrorOut"},"ErrorResponse":{"properties":{"status":{"type":"string","const":"error","title":"Status","default":"error"},"code":{"type":"string","title":"Code"},"message":{"type":"string","title":"Message"},"request_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Request Id"}},"type":"object","required":["code","message"],"title":"ErrorResponse","description":"Envelope lỗi HTTP (4xx/5xx) — khác với AgentResponse(type='error') là lỗi trong hội thoại."},"FieldChangeOut":{"properties":{"field":{"type":"string","title":"Field"},"label":{"type":"string","title":"Label"},"before":{"title":"Before"},"after":{"title":"After"}},"type":"object","required":["field","label"],"title":"FieldChangeOut"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LegacyChatRequest":{"properties":{"query":{"type":"string","maxLength":10000,"minLength":1,"title":"Query"},"session_id":{"type":"string","maxLength":100,"minLength":1,"title":"Session Id"},"lat":{"anyOf":[{"type":"number","maximum":90.0,"minimum":-90.0},{"type":"null"}],"title":"Lat"},"lng":{"anyOf":[{"type":"number","maximum":180.0,"minimum":-180.0},{"type":"null"}],"title":"Lng"},"image_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image Url"}},"type":"object","required":["query","session_id"],"title":"LegacyChatRequest"},"Location":{"properties":{"lat":{"type":"number","maximum":90.0,"minimum":-90.0,"title":"Lat"},"lng":{"type":"number","maximum":180.0,"minimum":-180.0,"title":"Lng"}},"type":"object","required":["lat","lng"],"title":"Location"},"MetaOut":{"properties":{"contract_version":{"type":"string","title":"Contract Version","default":"1.0"},"profile":{"type":"string","enum":["customer","management"],"title":"Profile"},"role":{"type":"string","enum":["guest","customer","supplier","admin"],"title":"Role"},"planner":{"type":"string","enum":["rules","llm"],"title":"Planner"},"tools_used":{"items":{"type":"string"},"type":"array","title":"Tools Used"},"data_source":{"type":"string","const":"backend","title":"Data Source","default":"backend"}},"type":"object","required":["profile","role","planner"],"title":"MetaOut"},"SourceOut":{"properties":{"system":{"type":"string","const":"backend","title":"System"},"resource":{"type":"string","title":"Resource"},"freshness":{"type":"string","enum":["realtime","reference","semantic"],"title":"Freshness"},"fetched_at":{"type":"string","format":"date-time","title":"Fetched At"},"record_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Record Id"},"version":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Version"}},"type":"object","required":["system","resource","freshness","fetched_at"],"title":"SourceOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}